fix: bug with sublayers
